Pulled out ENI discovery from EC2.  This can be enabled by enabling the 'eni' service in the configuration
